[SDAO-13] Add integration test using `tezos-client` bindings
Problem: Integration end-to-end testing is necessary to always ensure Lorentz
contract written works on real network. Previously it was done using
python scripts and this solution was very fast to prototype but not very
pleasant to use.

Solution: Implement test scenario based on `tezos-client` executable turtle
wrapper and integrate it into current test environment we are using.