[SDAO-23] Make multisig polymorphic with ability to rotate keys
Problem: Previous version of multisig coult work with STKR.Parameter
only. We want to be able to upgrade STKR while preserving Multisig
on chain. In order to achieve this, we need Multisig parameter type
to be polymorphic over the type of the callee. Moreover, we need to
be able to rotate keys of multisig and disable the ability to change
the owner of STKR.

Solution:
1. Add `Order` type that encapsulates both key rotation and generic
contract calls.
2. Remove `SetOperationsTeam` endpoint from STKR.
3. Remove `team` from STKR, make `owner` point to Multisig.