[SDAO-30] Add `EnsureOwner` newtype for STKR
Problem: currently the only check way that protected with multisig are actually
called using Multising proxy is instruction `ensureOwner` used in these
entrypoints. This approach is not very clear from the viewer perspective and
not reflected anywhere in `Parameter` type.

Solution: Make `newtype EnsureOwner a` and `ensureOwner` binding which checks
owner and unwraps from newtype if check was successful.